chore: Refactor file names and folder structure
